% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_supp_hcup_dat.R
\name{add_supp_hcup_dat}
\alias{add_supp_hcup_dat}
\title{combine additional HCUP data with data originally pulled from HCUP}
\usage{
add_supp_hcup_dat(base_dat, supp_dat)
}
\arguments{
\item{base_dat}{original processed data from hcup}

\item{supp_dat}{additional/supplementary processed data from HCUP}

\item{transfer}{transfer type of the hospitalization: not a transfer,}
}
\value{
Returns a tibble with the original data and the new data combined.
}
\description{
The original HCUP data took a while to run and later I decided
there was additional data I wanted to pull from HCUP so it was faster to
pull it separately and then join it based on the key. This function takes
the original data (\code{base_dat}) and joins the supplementary data
to it (\code{supp_dat}).
}
