% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjust_for_queue_l}
\alias{adjust_for_queue_l}
\alias{adjust_for_queue_i}
\title{Adjust admit queue list}
\usage{
adjust_for_queue_i(num_pat, q_n)
}
\arguments{
\item{patients}{List of patients at each facility}

\item{q_vec}{number in queue total}
}
\value{
adjusted queue list.
}
\description{
Check how many patients from each facility need to go to the queue.
}
