% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pat_mov_df.R
\name{create_pat_mov_df}
\alias{create_pat_mov_df}
\title{Construct data frame of patient moves for the day}
\usage{
create_pat_mov_df(ids, start_loc, drawn_moves)
}
\arguments{
\item{ids}{vector of patient ids}

\item{start_loc}{vector of patient starting room location (1/0) for (icu/non) for the day}

\item{drawn_moves}{vector of the number of icu/non moves for the day, designed to take
output from my \code{sample_day_mvts()} rcpp function}
}
\value{
Data frame with number of rows equal to the number of patient moves per day
}
\description{
Take all the patients in the hospital (their id, visit LOS, &
current room type icu/non) and construct a data frame of their movements for
the coming day. The input vectors should all be of the same length, with length
equal to the number of patients in the hospital on that day. If a patient's
drawn_moves for the day are "0,0" they will not be in the resulting data set.
}
