% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_CDI_los.R
\name{draw_CDI_los}
\alias{draw_CDI_los}
\title{draw CDI los based on current day during simulation}
\usage{
draw_CDI_los(
  dist = CDI_los_dist_mdc_tran6days,
  current_day,
  transfer,
  md_cat,
  hcup_los,
  hosp_id
)
}
\arguments{
\item{dist}{a list object with los distributions by mdc and transfer type}

\item{current_day}{day of hospitalization for the patient}

\item{transfer}{transfer type of the hospitalization: not a transfer,
transfer not last, transfer last}

\item{md_cat}{major diagnostic category, 0-25}

\item{hcup_los}{the patient's los in the HCUP data}

\item{hosp_id}{hospital the patient is from}
}
\value{
Returns a integer (or vector, function is vectorized) as the new LOS
}
\description{
If a patient gets CDI during the simulation, their LOS should
be adjusted to be longer. Draw a new LOS from the matching transfer/MDC
distribution truncated above the current LOS + 1. If the current_day is
great than or equal to the maximum los for the distribution, then the new LOS
is not drawn from the truncated distribution, instead it is choosen from the
maximum of the (current_day + 1) or los_hcup
}
