% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_day_mvts_cpp}
\alias{get_day_mvts_cpp}
\alias{combine_vectors}
\title{Determine the time/location for each patient's daily movements.}
\usage{
combine_vectors(x, y)
}
\arguments{
\item{ids}{vector of patients to get daily movement pattern for}

\item{icu_num}{vector of equal length to ids with the number of icu movements
for each patient}

\item{non_num}{vector of equal length to ids with the number of non-icu
movements for each patient}

\item{end_loc}{vector of equal length to ids, indicating whether each
patient ends in the icu or non-icu}
}
\value{
returns a data frame with one row per patient movement in the day.
Columns include patient id, room type (icu or non), and time block for
movement.
}
\description{
Check how many patients from each facility need to go to the queue.
}
