% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{move_rooms_cpp_seed}
\alias{move_rooms_cpp_seed}
\title{Move patient rooms}
\arguments{
\item{pat_rm_type}{Data frame of patients and the room type they are to go to.}

\item{icu}{vector of number of icu rooms available at each hospital}

\item{non}{vector of number of non-icu rooms available at each hospital}

\item{seed}{seed to be passed in to rcpp}
}
\value{
returns data frame with one column for patient and a second column
for the room number to which they are assigned.
}
\description{
Move patients from one room to another (similar to assign
rooms function). Designed to deal with one facility at a time.
}
