% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{num_in_q_by_risk}
\alias{num_in_q_by_risk}
\title{Adjust admit patient list for the number of patients going to the queue.}
\arguments{
\item{risks_beds}{Data frame with columns: hospital id, number of patients
at each risk level 1-4, number of icu beds, number of non-icu beds.}
}
\value{
List with length equal to the number of facilities with the new
number of patients at each risk level after sending the proper number of patients
to the queue.
}
\description{
Check how many patients from each facility need to go to the queue.
}
