% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_in_one_state.R
\name{pat_in_one_state}
\alias{pat_in_one_state}
\title{Verify patient(s) in one disease state}
\usage{
pat_in_one_state(pats)
}
\arguments{
\item{pats}{vector of patient ids of interest}
}
\value{
logical vector
}
\description{
Take a vector of patients IDs and check if patients are in
more than one state
}
