% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_assign_rooms_cpp_seed}
\alias{rcpp_assign_rooms_cpp_seed}
\title{Assign rooms to patients}
\arguments{
\item{pat_risks}{data frame where each row is a facility and four columns
one for each risk level to hold the number of patients at each risk level
per hospital}

\item{icu}{vector of number of icu rooms available at each hospital}

\item{non}{vector of number of non-icu rooms available at each hospital}

\item{seed}{seed to be passed in to rcpp}
}
\value{
returns data frame with one column for patient and a second column
for the room number to which they are assigned.
}
\description{
Randomly assign patients to hospital rooms with rcpp seed.
Designed to deal with one facility at a time.
}
