% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexptrunc.R
\name{rexptrunc}
\alias{rexptrunc}
\title{Truncated exponential distribution}
\usage{
rexptrunc(n, range, rate)
}
\arguments{
\item{n}{number of observations}

\item{range}{numeric vector of length two with bounds for the truncated distribution}

\item{rate}{rate for distribution}
}
\value{
Vector of length \code{n} of random samples from truncated exponential distribution
}
\description{
Random generation for the truncated exponential distribution
}
