% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_dis_state.R
\name{which_dis_state}
\alias{which_dis_state}
\title{check disease states from a vector of patient IDs}
\usage{
which_dis_state(pats)
}
\value{
character vector of disease states
}
\description{
take vector of patient IDS and return disease status of each patient
}
